Dataset: Health Metrics Data - G7 Countries

A subset of data from the IHME GBD on Deaths, DALYs, YLLs, YLDs, Incidence and Prevalence, age standardized all Causes and level 2 causes. For years 2010 and 2019.

Usage

g7_hmetrics

Format

A dataframe with 8610 rows and 9 variables:

The variables are as follows:

measure

character, metric name

location

character, country

sex

character, Female, Male, Both

cause

character, all causes, level 2 causes

year

integer, year

val

numeric, estimated values

upper

numeric, estimated upper values

lower

numeric, estimated lower values

Source

https://vizhub.healthdata.org/gbd-results/

Details

Locations available are Global, Canada, France, Germany, Italy, Japan, UK, and US.