Download, Unzip and Read Data: getunz

Download, Unzip and Read Data: getunz

Usage

getunz(url)

Arguments

url

A url string for a .zip file.

Value

A dataframe object from a zipped file. Particulary useful For downloading data from IHME GBD Results: "https://vizhub.healthdata.org/gbd-results/". The function takes the url, creates a temp directory, unzip the file, if more than one csv files is available, it lists the files, and reads them.

Select a dataset from the IHME GBD results and download it. You will receive an email with a url. Use the url to download the data.

Examples

if (FALSE) { # \dontrun{
url <- "https://www.healthdata.org/.../some-file.zip"
getunz(url)
} # }