hmsidwR: Health Metrics and the Spread of Infectious Diseases

A collection of data and functions used in the Health Metrics and the Spread of Infectious Diseases book. The book provides the tools for calculating the DALYs (YLLs+YLDs), and other important health metrics. It also shows the R code for modeling infectious diseases with machine learning. Finally, allows the user to understand how to visualize the results and making maps. Gazzelloni (2024) doi:10.5281/zenodo.10818338 .

See also

Author

Maintainer: Federica Gazzelloni fede.gazzelloni@gmail.com (ORCID)