lmforecast function

lmforecast function

Usage

lmforecast(x, a, m, e)

Arguments

x

A numeric vector, identifying the covariate of the model

a

A numeric vector, The intercept

m

A numeric vector, The slope, or the betas estimated values

e

A numeric vector, The error

Value

A dataframe containing the vector with the estimation, and the original data

Examples

x <- rnorm(n = 50, mean = 0, sd = 1)
lmforecast(
x = x,
a = 0.1,
m = dunif(x, min = 0.02, max = 1),
e = rnorm(length(x), mean = 0, sd = 1))
#>            pred            x       error
#> 1   0.962086482 -1.400043517  0.86208648
#> 2   0.117290867  0.255317055 -0.24323674
#> 3  -0.106087195 -2.437263611 -0.20608719
#> 4   0.119177592 -0.005571287  0.01917759
#> 5   0.763798225  0.621552721  0.02956075
#> 6   0.649827542  1.148411606  0.54982754
#> 7  -2.174114857 -1.821817661 -2.27411486
#> 8   2.782557184 -0.247325302  2.68255718
#> 9  -0.261221255 -0.244199607 -0.36122126
#> 10  0.313355750 -0.282705449  0.21335575
#> 11  1.174345882 -0.553699384  1.07434588
#> 12  0.076730162  0.628982042 -0.66508825
#> 13  1.213952419  2.065024895  1.11395242
#> 14 -0.145896412 -1.630989402 -0.24589641
#> 15 -0.554678666  0.512426950 -1.17756331
#> 16 -0.875850616 -1.863011492 -0.97585062
#> 17  1.165057320 -0.522012515  1.06505732
#> 18  0.231670635 -0.052601910  0.13167063
#> 19  1.142706710  0.542996343  0.48862881
#> 20 -1.599450568 -0.914074827 -1.69945057
#> 21 -0.893027714  0.468154420 -1.47073631
#> 22  0.754508768  0.362951256  0.28415034
#> 23  1.437320413 -1.304543545  1.33732041
#> 24  1.089529264  0.737776321  0.23669628
#> 25  1.418293384  1.888504929  1.31829338
#> 26  0.623909788 -0.097445104  0.52390979
#> 27  0.706748047 -0.935847354  0.60674805
#> 28 -0.009935672 -0.015950311 -0.10993567
#> 29  0.272181715 -0.826788954  0.17218172
#> 30  0.009672713 -1.512399651 -0.09032729
#> 31  2.978795576  0.935363190  1.92434334
#> 32  1.578483178  0.176488611  1.29839276
#> 33  1.097449905  0.243685465  0.74879127
#> 34  0.656224329  1.623548883  0.55622433
#> 35 -0.333932689  0.112038083 -0.54825726
#> 36  1.210534893 -0.133997013  1.11053489
#> 37 -2.512334333 -1.910087468 -2.61233433
#> 38 -0.055693776 -0.279237242 -0.15569378
#> 39  0.533889790 -0.313445978  0.43388979
#> 40 -0.281951112  1.067307879 -0.38195111
#> 41  0.595651708  0.070034850  0.42418757
#> 42  1.163101996 -0.639123324  1.06310200
#> 43  1.148712620 -0.049964899  1.04871262
#> 44  0.061897105 -0.251483443 -0.03810289
#> 45  1.040023527  0.444797116  0.48614892
#> 46  1.772882611  2.755417575  1.67288261
#> 47 -0.206880163  0.046531380 -0.35436116
#> 48  1.635846937  0.577709069  0.94634789
#> 49  1.537433371  0.118194874  1.31682636
#> 50 -0.196640025 -1.911720491 -0.29664002